Track::Plot: fix implicit cast

+++ Source/Track/AnlTrackPlot.cpp
@@ -149,7 +149,7 @@ void Track::Plot::paintMarkers(juce::Graphics& g, juce::Rectangle<float> const& 
             auto const w = Tools::secondsToPixel(end, timeRange, bounds) - x1;
             rectangles.addWithoutMerging({x1, clipBounds.getY(), std::max(w, 1.0f), clipBounds.getHeight()});
             
-            if(showLabel && !it->label.empty() && (labels.empty() || (std::get<1>(labels.back()) + std::get<2>(labels.back())) <= x2))
+            if(showLabel && !it->label.empty() && (labels.empty() || static_cast<float>(std::get<1>(labels.back()) + std::get<2>(labels.back())) <= x2))
             {
                 auto const text = juce::String(it->label) + unit;
                 auto const textWidth = font.getStringWidth(text) + 2;
